requirements.txt
category-encoders 
certifi 
charset-normalizer
contourpy 
cycler 
fonttools 
idna 
imbalanced-learn 
imblearn 
joblib 
kiwisolver 
matplotlib 
numpy 
packaging 
pandas 
patsy 
pillow 
pyparsing 
python-dateutil 
pytz 
requests 
scikit-learn 
scipy 
seaborn 
six 
statsmodels
threadpoolctl 
tzdata 
ucimlrepo 
urllib3 
xgboost 
matplotlib-inline
streamlit 
plotly 
PyJWT
fastapi
uvicorn 
jinja2
python-multipart
psutil
cloudpickle
streamlit-navigation-bar

